requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
redis==5.0.1
rq==1.16.0
ultralytics==8.0.209
opencv-python==4.8.1.78
numpy==1.24.3
opencv-contrib-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
pydantic==2.4.2
pydantic-settings==2.0.3
aiofiles==23.2.1
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pillow==10.1.0
